+++ src/query/auth.ts
@@ -7,7 +7,7 @@ import { LoginForm } from 'model/auth';
 import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useErrorMessageStore } from 'store/errorMessageStore';
-import useUserTypeStore from 'store/userType';
+import useUserTypeStore from 'store/useUserTypeStore';
 import useUserStore from 'store/user';
 import { isKoinError } from '@bcsdlab/koin';
